backend/requirements.txt
asgiref==3.8.1
boto3==1.37.26
botocore==1.37.26
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
Django==4.2.20
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.16.0
jmespath==1.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
